Updates the Categories API to v3
Adds the ability to create a new Category.
Note: the API for find sub-categories has changed from a single
integer with the parentId to a Hash with the `parent: id` key.
This is so the API can be used with other keys as well (like
whether or not to return hidden Categories)