Fixes infinite loop with an incorrect total.
Apparently the total that comes back in the response from Ecwid
is higher than it should be. This caused an infinite loop because
`count + offset` would never reach the total. Eventually, the offset
would be high enough that there weren't any results. So we'll exit
the loop when the `count == 0` as well.