Refactors image uploading and raising exeptions
Problems that bugged me:

* the `Api` mixin module that shared the `#raise_on_failure` method
* the duplication of how `#raise_on_failure` was being used
* the duplication of code for uploading images

I realized that `#raise_on_failure` was only being used for requests
that changed state: post, put and delete. It seemed like it would
be better suited for the `Client` rather than implementing it
in `Api`s and `Entities`.

This change ensures that our `ResponseErrors` are always raised
for failed post, put and delete requests. It also rendered the `Api`
mixin useless. Yay!

I then realized that the duplicated code for uploading images could
easily be factored into the `Client` as well. Thus the
`Client#post_image` helper method was born.

Now code that uses the `Client` is much cleaner and consistent.